requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.9
torch==2.1.2
torchvision==0.16.2
Pillow>=11.0.0,<12.0.0
numpy<2
timm
matplotlib
seaborn
scikit-learn
pandas
streamlit>=1.31.0
